Condense WorkflowHandle and status
