+++ dbos/__init__.py
@@ -2,11 +2,11 @@ from . import error as error
 from .context import SetWorkflowUUID as SetWorkflowUUID
 from .dbos import DBOS as DBOS
 from .dbos import DBOSConfiguredInstance as DBOSConfiguredInstance
+from .dbos import WorkflowHandle as WorkflowHandle
+from .dbos import WorkflowStatus as WorkflowStatus
 from .dbos_config import ConfigFile as ConfigFile
 from .dbos_config import load_config
 from .system_database import WorkflowStatusString as WorkflowStatusString
-from .workflow import WorkflowHandle as WorkflowHandle
-from .workflow import WorkflowStatus as WorkflowStatus
 
 __all__ = [
     "ConfigFile",
